Datafest Africa ML and EDA Hackathon
This case study aims analyze the status of applicants and their loan application in order to automate the process by building a machine learning model to predict the outcome of the lending process if the credit facility process was completed via e-process, that is E-Signed or not. Recognition of such applicants using Exploratory Data Analysis (EDA) techniques is the main focus of this case study.